% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_functions.R
\name{plot.SCINA}
\alias{plot.SCINA}
\title{A function to plot SCINA results in a heatmap.}
\usage{
\method{plot}{SCINA}(exp, results, signatures)
}
\arguments{
\item{exp}{See more details in \code{\link{SCINA}}}

\item{results}{An output object returned from SCINA.}

\item{signatures}{See more details in \code{\link{SCINA}}}
}
\value{
Plot a heatmap showing signature genes' expression level and SCINA predicted cell types.
}
\description{
A function to plot SCINA results in a heatmap.
}
\examples{
plot.SCINA(exp, results, signatures)

}
