% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_functions.R
\name{preprocess.signatures}
\alias{preprocess.signatures}
\title{A function to convert signatures uploaded via .csv files to lists used by SCINA.}
\usage{
preprocess.signatures(file_path)
}
\arguments{
\item{file_path}{The path of the .csv file. The first row of the file should be cell type names. Each column is occupied by the signature genes/protein markers for the cell type in the first row. Please find more details in \code{\link{SCINA}}.}
}
\value{
A list of signature gene lists as an input for SCINA.
}
\description{
A function to convert signatures uploaded via .csv files to lists used by SCINA.
}
\examples{
signatures = preprocess.signatures("./example/example_signatures.csv")
}
